fixed test

+++ src/main/java/org/example/individualbackend/config/db/conrs/security/token/AccessToken.java
@@ -5,7 +5,7 @@ import java.util.Set;
 public interface AccessToken {
     String getSubject();
 
-    Long getUserId();
+    Integer getUserId();
 
     Set<String> getRoles();
